refactor: extract text area resizer
